requirement.txt
pycrypto-2.6.1
flask-1.1.2
flask-cors-3.0.8
requests-2.24.0